Rename package msgq → queue
